Struct tokio_netstring::Builder

pub struct Builder { /* fields omitted */ }

Configure netstring delimited FramedRead, FramedWrite, and Framed values.

Builder enables constructing configured netstring delimited framers. Note that not all configuration settings apply to both encoding and decoding. See the documentation for specific methods for more detail.

Methods

impl Builder

fn new() -> Builder

Creates a new length delimited framer builder with default configuration values.

Examples

use tokio_netstring::Builder;

Builder::new()
    .length_field_offset(0)
    .strip_frame(true)
    .new_read(io);

fn max_frame_length(&mut self, val: usize) -> &mut Self

Sets the max frame length

This configuration option applies to both encoding and decoding. The default value is 8MB.

When decoding, the length field read from the byte stream is checked against this setting before any adjustments are applied. When encoding, the length of the submitted payload is checked against this setting.

Examples

use tokio_netstring::Builder;

Builder::new()
    .max_frame_length(8 * 1024)
    .new_read(io);

fn length_field_offset(&mut self, val: usize) -> &mut Self

Sets the number of bytes in the header before the length field

This configuration option only applies to decoding.

Examples

use tokio_netstring::Builder;

Builder::new()
    .length_field_offset(1)
    .new_read(io);

fn strip_frame(&mut self, val: bool) -> &mut Self

Sets wether or not to streip the frame and leave only the payload

Default value is true

This configuration option only applies to decoding

Examples

use tokio_netstring::Builder;

Builder::new()
    .strip_frame(false)
    .new_read(io);

fn new_read<T>(&self, upstream: T) -> FramedRead<T> where T: AsyncRead

Create a configured length delimited FramedRead

Examples

use tokio_netstring::Builder;

Builder::new()
    .length_field_offset(0)
    .strip_frame(false)
    .new_read(io);

fn new_write<T, B>(&self, inner: T) -> FramedWrite<T, B> where T: AsyncWrite, B: IntoBuf

Create a configured length delimited FramedWrite

Examples

netstring::Builder::new()
    .new_write(io);

fn new_framed<T, B>(&self, inner: T) -> Framed<T, B> where T: AsyncRead + AsyncWrite, B: IntoBuf

Create a configured netstring delimited Framed

Examples

netstring::Builder::new()
    .new_framed(io);